+++ Src/Private/Get-AbrDomainSection.ps1
@@ -88,25 +88,25 @@ function Get-AbrDomainSection {
                                                     }
                                                 }
                                             }
-                                            if ($HealthCheck.DomainController.Diagnostic) {
-                                                try {
-                                                    $DCDiagObj = foreach ($DC in $DCs) {
-                                                        if (Get-DCWinRMState -ComputerName $DC -DCStatus ([ref]$DCStatus)) {
-                                                            Get-AbrADDCDiag -Domain $Domain -DC $DC
-                                                        }
-                                                    }
-                                                    if ($DCDiagObj) {
-                                                        Section -Style Heading4 'DC Diagnostic' {
-                                                            Paragraph "The following section provides a summary of the Active Directory DC Diagnostic."
-                                                            BlankLine
-                                                            $DCDiagObj
-                                                        }
-                                                    }
-                                                } catch {
-                                                    Write-PScriboMessage -IsWarning "Error: Connecting to remote server $DC failed: WinRM cannot complete the operation. ('DCDiag Information)"
-                                                    Write-PScriboMessage -IsWarning $_.Exception.Message
-                                                }
-                                            }
+                                            # if ($HealthCheck.DomainController.Diagnostic) {
+                                            #     try {
+                                            #         $DCDiagObj = foreach ($DC in $DCs) {
+                                            #             if (Get-DCWinRMState -ComputerName $DC -DCStatus ([ref]$DCStatus)) {
+                                            #                 Get-AbrADDCDiag -Domain $Domain -DC $DC
+                                            #             }
+                                            #         }
+                                            #         if ($DCDiagObj) {
+                                            #             Section -Style Heading4 'DC Diagnostic' {
+                                            #                 Paragraph "The following section provides a summary of the Active Directory DC Diagnostic."
+                                            #                 BlankLine
+                                            #                 $DCDiagObj
+                                            #             }
+                                            #         }
+                                            #     } catch {
+                                            #         Write-PScriboMessage -IsWarning "Error: Connecting to remote server $DC failed: WinRM cannot complete the operation. ('DCDiag Information)"
+                                            #         Write-PScriboMessage -IsWarning $_.Exception.Message
+                                            #     }
+                                            # }
                                             try {
                                                 $ADInfraServices = foreach ($DC in $DCs) {
                                                     if (Get-DCWinRMState -ComputerName $DC -DCStatus ([ref]$DCStatus)) {